[Refactor] reorganize components
