+++ app/controllers/conference_rooms_controller.rb
@@ -12,7 +12,7 @@ class ConferenceRoomsController < ApplicationController
   private
 
   def create_props
-    @props = {conference_room: conference_room}.compact
+    @props = {conference_room: conference_room}
   end
 
   def conference_room
