+++ client/app/Register.jsx
@@ -1,5 +1,5 @@
 import ReactOnRails from 'react-on-rails';
-import CalendarRoot from './components/ReactRoot';
+import CalendarRoot from './components/calendar/ReactRoot';
 import ConferenceRoomRoot from './components/conference_room/ReactRoot';
 
 import './app.scss';
