+++ client/app/components/calendar/ReactRoot.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Router, Route, browserHistory } from 'react-router';
 
-import App from './App';
+import CalendarContainer from './CalendarContainer';
 
 const wrapComponent = (Component, ownProps) => (props) => (
   <Component {...ownProps} {...props} />
@@ -9,6 +9,6 @@ const wrapComponent = (Component, ownProps) => (props) => (
 
 export default (props) => (
   <Router history={browserHistory}>
-    <Route path="/(:date)" component={wrapComponent(App, props)} />
+    <Route path="/(:date)" component={wrapComponent(CalendarContainer, props)} />
   </Router>
 );
