+++ client/app/components/calendar/app.scss
