+++ client/app/components/calendar/layout/refresh_button.scss
