+++ client/app/components/calendar/table/calendar.scss
