+++ client/test/components/calendar/Calendar.spec.jsx
@@ -6,9 +6,9 @@ import { shallow, mount } from 'enzyme';
 import { expect } from 'chai';
 import DefaultProps from '../../factories/DefaultProps';
 import { Table } from 'react-bootstrap';
-import Calendar from 'components/calendar/Calendar';
-import CalendarHeader from 'components/calendar/CalendarHeader';
-import CalendarRow from 'components/calendar/CalendarRow';
+import Calendar from 'components/calendar/table/Calendar';
+import CalendarHeader from 'components/calendar/table/CalendarHeader';
+import CalendarRow from 'components/calendar/table/CalendarRow';
 import RoomFilters from 'components/calendar/filters/RoomFilters';
 import * as FiltersHelper from 'helpers/FiltersHelper';
 import proxyquire from 'proxyquire';
