+++ client/test/components/calendar/CalendarRow.spec.jsx
@@ -2,9 +2,9 @@ import moment from 'moment';
 import React from 'react';
 import { shallow } from 'enzyme';
 import sinon from 'sinon';
-import CalendarRow from 'components/calendar/CalendarRow';
+import CalendarRow from 'components/calendar/table/CalendarRow';
 import EventWrapper from 'components/calendar/event/EventWrapper';
-import TimeCell from 'components/calendar/TimeCell';
+import TimeCell from 'components/calendar/table/TimeCell';
 import { expect } from 'chai';
 
 describe('<CalendarRow />', () => {
