+++ client/test/components/calendar/modal/CreateEventModal.spec.jsx
@@ -3,11 +3,11 @@ import { Modal } from 'react-bootstrap';
 import { shallow } from 'enzyme';
 import { expect } from 'chai';
 import sinon from 'sinon';
-import CreateEventModal from 'components/modal/CreateEventModal';
+import CreateEventModal from 'components/calendar/modal/CreateEventModal';
 import ConferenceRoom from 'test/factories/ConferenceRoom';
-import ModalHeader from 'components/modal/layout/ModalHeader';
-import ModalBody from 'components/modal/layout/ModalBody';
-import ModalFooter from 'components/modal/layout/ModalFooter';
+import ModalHeader from 'components/calendar/modal/layout/ModalHeader';
+import ModalBody from 'components/calendar/modal/layout/ModalBody';
+import ModalFooter from 'components/calendar/modal/layout/ModalFooter';
 
 describe('<CreateEventModal />', () => {
   const props = {
