+++ client/test/components/calendar/modal/layout/ErrorField.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Alert } from 'react-bootstrap';
 import { mount } from 'enzyme';
 import { expect } from 'chai';
-import ErrorField from 'components/modal/layout/ErrorField';
+import ErrorField from 'components/calendar/modal/layout/ErrorField';
 
 describe('<ErrorField />', () => {
   it('does not render <Alert /> by default', () => {
