+++ client/test/components/calendar/modal/layout/body/RoomSelectGroup.spec.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
 import { expect } from 'chai';
-import RoomSelectGroup from 'components/modal/layout/body/RoomSelectGroup';
+import RoomSelectGroup from 'components/calendar/modal/layout/body/RoomSelectGroup';
 import ConferenceRoom from 'test/factories/ConferenceRoom';
 
 describe('<RoomSelectGroup />', () => {
